re-structure project files
