+++ src/components/common/Button.css
@@ -1,0 +1,94 @@
+/* Button Component Styles */
+.btn {
+	display: inline-flex;
+	align-items: center;
+	justify-content: center;
+	gap: var(--spacing-sm);
+	border: none;
+	border-radius: var(--radius-xl);
+	font-weight: var(--font-semibold);
+	cursor: pointer;
+	transition: all var(--transition-normal);
+	text-decoration: none;
+	font-family: inherit;
+	white-space: nowrap;
+}
+
+.btn:disabled {
+	opacity: 0.6;
+	cursor: not-allowed;
+	pointer-events: none;
+}
+
+/* Button Sizes */
+.btn--small {
+	padding: var(--spacing-sm) var(--spacing-md);
+	font-size: var(--text-sm);
+}
+
+.btn--medium {
+	padding: var(--spacing-md) var(--spacing-xl);
+	font-size: var(--text-base);
+}
+
+.btn--large {
+	padding: var(--spacing-lg) var(--spacing-2xl);
+	font-size: var(--text-lg);
+}
+
+/* Button Variants */
+.btn--primary {
+	background: var(--primary-gradient);
+	color: var(--text-white);
+	box-shadow: var(--shadow-primary);
+}
+
+.btn--primary:hover {
+	transform: translateY(-2px);
+	box-shadow: var(--shadow-primary-lg);
+}
+
+.btn--secondary {
+	background: transparent;
+	color: var(--primary-color);
+	border: 2px solid var(--primary-color);
+}
+
+.btn--secondary:hover {
+	background: var(--primary-color);
+	color: var(--text-white);
+	transform: translateY(-2px);
+}
+
+.btn--outline {
+	background: transparent;
+	color: var(--text-primary);
+	border: 2px solid #ddd;
+}
+
+.btn--outline:hover {
+	border-color: var(--primary-color);
+	color: var(--primary-color);
+}
+
+.btn--ghost {
+	background: transparent;
+	color: var(--text-secondary);
+	border: 1px solid rgba(255, 87, 87, 0.3);
+}
+
+.btn--ghost:hover {
+	background: rgba(255, 87, 87, 0.1);
+	border-color: var(--primary-color);
+	color: var(--primary-color);
+}
+
+/* Button States */
+.btn:focus {
+	outline: none;
+	box-shadow: 0 0 0 3px rgba(255, 87, 87, 0.3);
+}
+
+.btn:active {
+	transform: translateY(0);
+}
