+++ src/components/common/Card.css
@@ -1,0 +1,74 @@
+/* Card Component Styles */
+.card {
+	background: var(--bg-white);
+	border-radius: var(--radius-xl);
+	box-shadow: var(--shadow-lg);
+	transition: all var(--transition-normal);
+	overflow: hidden;
+	border: 1px solid rgba(255, 87, 87, 0.1);
+}
+
+.card--hover:hover {
+	transform: translateY(-5px);
+	box-shadow: var(--shadow-xl);
+	border-color: rgba(255, 87, 87, 0.3);
+}
+
+/* Card Variants */
+.card--default {
+	background: var(--bg-white);
+}
+
+.card--light {
+	background: var(--bg-light);
+}
+
+.card--gradient {
+	background: var(--primary-gradient);
+	color: var(--text-white);
+}
+
+.card--glass {
+	background: rgba(255, 255, 255, 0.9);
+	backdrop-filter: blur(10px);
+	border: 1px solid rgba(255, 87, 87, 0.1);
+}
+
+/* Card Sections */
+.card__header {
+	padding: var(--spacing-xl) var(--spacing-xl) 0;
+}
+
+.card__body {
+	padding: var(--spacing-xl);
+}
+
+.card__footer {
+	padding: 0 var(--spacing-xl) var(--spacing-xl);
+}
+
+/* Card with Image */
+.card__image {
+	width: 100%;
+	height: 200px;
+	object-fit: cover;
+}
+
+/* Card Content */
+.card__title {
+	font-size: var(--text-xl);
+	font-weight: var(--font-bold);
+	color: var(--text-primary);
+	margin-bottom: var(--spacing-md);
+}
+
+.card__description {
+	color: var(--text-secondary);
+	line-height: 1.6;
+	margin-bottom: var(--spacing-lg);
+}
+
+.card--gradient .card__title,
+.card--gradient .card__description {
+	color: var(--text-white);
+}
