+++ src/components/common/index.js
@@ -1,0 +1,4 @@
+// Common components exports
+export { default as Button } from "./Button";
+export { default as Card } from "./Card";
+export { default as SectionTitle } from "./SectionTitle";
