+++ src/components/layout/Navbar.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Navbar.css";
+import "../../styles/layout/Navbar.css";
 
 const Navbar = () => {
 	return (
