+++ src/hooks/index.js
@@ -1,0 +1,46 @@
+import { useState, useEffect } from "react";
+
+// Custom hook for managing local storage
+export const useLocalStorage = (key, initialValue) => {
+	const [storedValue, setStoredValue] = useState(() => {
+		try {
+			const item = window.localStorage.getItem(key);
+			return item ? JSON.parse(item) : initialValue;
+		} catch (error) {
+			console.log(error);
+			return initialValue;
+		}
+	});
+
+	const setValue = (value) => {
+		try {
+			const valueToStore = value instanceof Function ? value(storedValue) : value;
+			setStoredValue(valueToStore);
+			window.localStorage.setItem(key, JSON.stringify(valueToStore));
+		} catch (error) {
+			console.log(error);
+		}
+	};
+
+	return [storedValue, setValue];
+};
+
+// Custom hook for detecting mobile devices
+export const useIsMobile = () => {
+	const [isMobile, setIsMobile] = useState(false);
+
+	useEffect(() => {
+		const checkIsMobile = () => {
+			setIsMobile(window.innerWidth <= 768);
+		};
+
+		checkIsMobile();
+		window.addEventListener("resize", checkIsMobile);
+
+		return () => {
+			window.removeEventListener("resize", checkIsMobile);
+		};
+	}, []);
+
+	return isMobile;
+};
