+++ src/styles/layout/Footer.css
@@ -1,0 +1,128 @@
+/* Footer Styles */
+.footer {
+	background: #333;
+	color: var(--text-white);
+	padding: var(--spacing-2xl) 0 var(--spacing-md);
+}
+
+.footer-content {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+	gap: var(--spacing-2xl);
+	margin-bottom: var(--spacing-xl);
+}
+
+.footer-section h4 {
+	margin-bottom: var(--spacing-md);
+	color: var(--primary-color-light);
+}
+
+.footer-section p {
+	color: #ccc;
+	line-height: 1.6;
+	margin-bottom: var(--spacing-md);
+}
+
+.social-icons {
+	display: flex;
+	gap: var(--spacing-md);
+	font-size: var(--text-2xl);
+}
+
+.social-icons span {
+	cursor: pointer;
+	transition: transform var(--transition-normal);
+}
+
+.social-icons span:hover {
+	transform: scale(1.2);
+}
+
+.footer-section ul {
+	list-style: none;
+}
+
+.footer-section ul li {
+	margin-bottom: var(--spacing-sm);
+}
+
+.footer-section ul li a {
+	color: #ccc;
+	text-decoration: none;
+	transition: color var(--transition-normal);
+}
+
+.footer-section ul li a:hover {
+	color: var(--primary-color-light);
+}
+
+.newsletter-signup {
+	display: flex;
+	gap: var(--spacing-sm);
+	margin-top: var(--spacing-md);
+}
+
+.newsletter-signup input {
+	flex: 1;
+	padding: var(--spacing-sm);
+	border: none;
+	border-radius: var(--radius-md);
+	background: #444;
+	color: var(--text-white);
+}
+
+.newsletter-signup input::placeholder {
+	color: #ccc;
+}
+
+.newsletter-signup button {
+	background: var(--primary-gradient);
+	color: var(--text-white);
+	border: none;
+	padding: var(--spacing-sm) var(--spacing-lg);
+	border-radius: var(--radius-md);
+	cursor: pointer;
+	font-weight: var(--font-semibold);
+}
+
+.footer-bottom {
+	border-top: 1px solid #555;
+	padding-top: var(--spacing-xl);
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	flex-wrap: wrap;
+	gap: var(--spacing-md);
+}
+
+.sponsors {
+	display: flex;
+	align-items: center;
+	gap: var(--spacing-md);
+}
+
+.sponsor-badges {
+	display: flex;
+	gap: var(--spacing-sm);
+}
+
+.sponsor-badge {
+	background: var(--primary-color);
+	color: var(--text-white);
+	padding: var(--spacing-xs) var(--spacing-sm);
+	border-radius: var(--radius-md);
+	font-size: var(--text-xs);
+	font-weight: var(--font-semibold);
+}
+
+/* Responsive Footer */
+@media (max-width: 768px) {
+	.footer-bottom {
+		flex-direction: column;
+		text-align: center;
+	}
+
+	.newsletter-signup {
+		flex-direction: column;
+	}
+}
