+++ src/styles/sections/CommunitySection.css
@@ -1,0 +1,164 @@
+/* Community Section Styles */
+.community-section {
+	padding: var(--spacing-4xl) 0;
+	background: var(--bg-white);
+}
+
+.community-tabs {
+	display: flex;
+	justify-content: center;
+	gap: var(--spacing-md);
+	margin-bottom: var(--spacing-2xl);
+}
+
+.community-tab {
+	padding: var(--spacing-md) var(--spacing-xl);
+	border: 2px solid var(--primary-color);
+	background: transparent;
+	color: var(--primary-color);
+	border-radius: var(--radius-xl);
+	cursor: pointer;
+	transition: all var(--transition-normal);
+	font-weight: var(--font-semibold);
+}
+
+.community-tab.active,
+.community-tab:hover {
+	background: var(--primary-color);
+	color: var(--text-white);
+}
+
+.marketplace-categories {
+	display: flex;
+	justify-content: center;
+	gap: var(--spacing-md);
+	margin-bottom: var(--spacing-xl);
+	flex-wrap: wrap;
+}
+
+.category-btn {
+	padding: var(--spacing-sm) var(--spacing-lg);
+	border: 2px solid #ddd;
+	background: var(--bg-white);
+	color: var(--text-secondary);
+	border-radius: var(--radius-xl);
+	cursor: pointer;
+	transition: all var(--transition-normal);
+	font-weight: var(--font-semibold);
+}
+
+.category-btn:hover {
+	border-color: var(--primary-color);
+	color: var(--primary-color);
+}
+
+.marketplace-grid {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+	gap: var(--spacing-xl);
+	margin-bottom: var(--spacing-3xl);
+}
+
+.product-card {
+	background: var(--bg-white);
+	border-radius: var(--radius-xl);
+	padding: var(--spacing-xl);
+	box-shadow: var(--shadow-lg);
+	transition: all var(--transition-normal);
+	text-align: center;
+}
+
+.product-card:hover {
+	transform: translateY(-5px);
+	box-shadow: var(--shadow-xl);
+}
+
+.product-image {
+	font-size: 4rem;
+	margin-bottom: var(--spacing-md);
+}
+
+.product-card h4 {
+	margin-bottom: var(--spacing-md);
+	color: var(--text-primary);
+}
+
+.seller-info {
+	margin-bottom: var(--spacing-md);
+}
+
+.seller-info span {
+	color: var(--text-secondary);
+	font-size: var(--text-sm);
+}
+
+.rating {
+	margin-top: var(--spacing-sm);
+}
+
+.price {
+	font-size: var(--text-2xl);
+	font-weight: var(--font-black);
+	color: var(--primary-color);
+}
+
+.community-stories {
+	margin-top: var(--spacing-3xl);
+}
+
+.community-stories h3 {
+	text-align: center;
+	margin-bottom: var(--spacing-xl);
+	color: var(--text-primary);
+}
+
+.stories-gallery {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(350px, 1fr));
+	gap: var(--spacing-xl);
+}
+
+.story-card {
+	background: linear-gradient(135deg, var(--bg-light), var(--bg-white));
+	border-radius: var(--radius-xl);
+	padding: var(--spacing-xl);
+	text-align: center;
+	transition: all var(--transition-normal);
+}
+
+.story-card:hover {
+	transform: translateY(-5px);
+	box-shadow: var(--shadow-lg);
+}
+
+.story-image {
+	font-size: 4rem;
+	margin-bottom: var(--spacing-md);
+}
+
+.story-card h4 {
+	margin-bottom: var(--spacing-md);
+	color: var(--text-primary);
+}
+
+.story-card p {
+	color: var(--text-secondary);
+	margin-bottom: var(--spacing-lg);
+	line-height: 1.6;
+}
+
+.story-btn {
+	background: var(--primary-gradient);
+	color: var(--text-white);
+	border: none;
+	padding: var(--spacing-sm) var(--spacing-xl);
+	border-radius: var(--radius-xl);
+	cursor: pointer;
+	transition: all var(--transition-normal);
+	font-weight: var(--font-semibold);
+}
+
+.story-btn:hover {
+	transform: translateY(-2px);
+	box-shadow: var(--shadow-primary);
+}
