+++ src/styles/sections/CompetitionsSection.css
@@ -1,0 +1,138 @@
+/* Competitions Section Styles */
+.competitions-section {
+	padding: var(--spacing-4xl) 0;
+	background: linear-gradient(135deg, var(--bg-light), var(--bg-white));
+}
+
+.competition-banners {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+	gap: var(--spacing-xl);
+	margin-bottom: var(--spacing-3xl);
+}
+
+.competition-banner {
+	background: var(--bg-white);
+	border-radius: var(--radius-xl);
+	padding: var(--spacing-2xl);
+	text-align: center;
+	box-shadow: var(--shadow-lg);
+	transition: all var(--transition-normal);
+}
+
+.competition-banner.featured {
+	background: var(--primary-gradient);
+	color: var(--text-white);
+}
+
+.competition-banner:hover {
+	transform: translateY(-5px);
+	box-shadow: var(--shadow-xl);
+}
+
+.competition-banner h3 {
+	margin-bottom: var(--spacing-md);
+	font-size: var(--text-2xl);
+}
+
+.competition-banner p {
+	margin-bottom: var(--spacing-xl);
+	line-height: 1.6;
+}
+
+.enter-btn {
+	background: transparent;
+	color: var(--primary-color);
+	border: 2px solid var(--primary-color);
+	padding: var(--spacing-md) var(--spacing-xl);
+	border-radius: var(--radius-xl);
+	cursor: pointer;
+	transition: all var(--transition-normal);
+	font-weight: var(--font-semibold);
+}
+
+.competition-banner.featured .enter-btn {
+	color: var(--text-white);
+	border-color: var(--text-white);
+}
+
+.enter-btn:hover {
+	background: var(--primary-color);
+	color: var(--text-white);
+}
+
+.competition-banner.featured .enter-btn:hover {
+	background: var(--text-white);
+	color: var(--primary-color);
+}
+
+.social-campaigns h3 {
+	text-align: center;
+	margin-bottom: var(--spacing-xl);
+	color: var(--text-primary);
+}
+
+.hashtag-campaigns {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+	gap: var(--spacing-xl);
+	margin-bottom: var(--spacing-2xl);
+}
+
+.hashtag-card {
+	background: var(--bg-white);
+	border-radius: var(--radius-xl);
+	padding: var(--spacing-xl);
+	text-align: center;
+	box-shadow: var(--shadow-lg);
+	transition: all var(--transition-normal);
+}
+
+.hashtag-card:hover {
+	transform: translateY(-5px);
+}
+
+.hashtag-card h4 {
+	color: var(--primary-color);
+	margin-bottom: var(--spacing-md);
+	font-size: var(--text-xl);
+}
+
+.hashtag-card p {
+	color: var(--text-secondary);
+	margin-bottom: var(--spacing-md);
+}
+
+.social-stats {
+	display: flex;
+	justify-content: center;
+	gap: var(--spacing-md);
+	font-size: var(--text-sm);
+	color: var(--text-secondary);
+}
+
+.embedded-content {
+	display: grid;
+	grid-template-columns: 1fr 1fr;
+	gap: var(--spacing-xl);
+}
+
+.embed-placeholder {
+	background: linear-gradient(135deg, var(--bg-light), #e9ecef);
+	border-radius: var(--radius-xl);
+	padding: var(--spacing-2xl);
+	text-align: center;
+	color: var(--text-secondary);
+	font-size: var(--text-xl);
+	min-height: 300px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+/* Responsive Competitions Section */
+@media (max-width: 768px) {
+	.embedded-content {
+		grid-template-columns: 1fr;
+	}
+}
