+++ src/styles/sections/EventsSection.css
@@ -1,0 +1,156 @@
+/* Events Section Styles */
+.events-section {
+	padding: var(--spacing-4xl) 0;
+	background: linear-gradient(135deg, var(--bg-lighter), var(--bg-light));
+}
+
+.events-interface {
+	display: grid;
+	grid-template-columns: 1fr 1fr;
+	gap: var(--spacing-2xl);
+	margin-top: var(--spacing-2xl);
+}
+
+.events-calendar,
+.live-map {
+	background: var(--bg-white);
+	border-radius: var(--radius-xl);
+	padding: var(--spacing-xl);
+	box-shadow: var(--shadow-lg);
+}
+
+.events-calendar h3,
+.live-map h3 {
+	margin-bottom: var(--spacing-xl);
+	color: var(--text-primary);
+}
+
+.event-item {
+	display: flex;
+	gap: var(--spacing-md);
+	padding: var(--spacing-lg);
+	border-radius: var(--radius-md);
+	background: var(--bg-light);
+	margin-bottom: var(--spacing-md);
+	transition: all var(--transition-normal);
+}
+
+.event-item:hover {
+	background: var(--bg-white);
+	box-shadow: var(--shadow-md);
+}
+
+.event-date {
+	background: var(--primary-gradient);
+	color: var(--text-white);
+	padding: var(--spacing-md);
+	border-radius: var(--radius-md);
+	text-align: center;
+	font-weight: var(--font-semibold);
+	min-width: 80px;
+}
+
+.event-info h4 {
+	margin-bottom: var(--spacing-sm);
+	color: var(--text-primary);
+}
+
+.event-info p {
+	color: var(--text-secondary);
+	margin-bottom: var(--spacing-md);
+}
+
+.event-actions {
+	display: flex;
+	gap: var(--spacing-sm);
+}
+
+.notify-btn,
+.interested-btn {
+	padding: var(--spacing-sm) var(--spacing-md);
+	border: none;
+	border-radius: var(--radius-md);
+	cursor: pointer;
+	font-size: var(--text-sm);
+	font-weight: var(--font-semibold);
+	transition: all var(--transition-normal);
+}
+
+.notify-btn {
+	background: var(--bg-white);
+	color: var(--primary-color);
+	border: 2px solid var(--primary-color);
+}
+
+.interested-btn {
+	background: var(--primary-color);
+	color: var(--text-white);
+}
+
+.notify-btn:hover,
+.interested-btn:hover {
+	transform: translateY(-2px);
+	box-shadow: var(--shadow-primary);
+}
+
+.map-container {
+	background: linear-gradient(135deg, var(--bg-light), #e9ecef);
+	border-radius: var(--radius-md);
+	padding: var(--spacing-xl);
+	min-height: 300px;
+	position: relative;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+.live-stream-marker {
+	position: absolute;
+	top: 20px;
+	right: 20px;
+	background: var(--primary-color);
+	color: var(--text-white);
+	padding: var(--spacing-sm) var(--spacing-md);
+	border-radius: var(--radius-md);
+	font-size: var(--text-sm);
+	animation: pulse 2s infinite;
+}
+
+.popup {
+	position: absolute;
+	bottom: 20px;
+	left: 20px;
+	background: var(--bg-white);
+	padding: var(--spacing-md);
+	border-radius: var(--radius-md);
+	box-shadow: var(--shadow-lg);
+	min-width: 200px;
+}
+
+.popup h4 {
+	margin-bottom: var(--spacing-sm);
+	color: var(--text-primary);
+}
+
+.popup p {
+	margin-bottom: var(--spacing-sm);
+	color: var(--text-secondary);
+	font-size: var(--text-sm);
+}
+
+.popup button {
+	background: var(--primary-color);
+	color: var(--text-white);
+	border: none;
+	padding: var(--spacing-sm) var(--spacing-md);
+	border-radius: var(--radius-sm);
+	cursor: pointer;
+	font-weight: var(--font-semibold);
+}
+
+/* Responsive Events Section */
+@media (max-width: 768px) {
+	.events-interface {
+		grid-template-columns: 1fr;
+	}
+}
