+++ src/styles/sections/HeroSection.css
@@ -1,0 +1,122 @@
+/* Hero Section Styles */
+.hero-section {
+	position: relative;
+	height: 100vh;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	overflow: hidden;
+}
+
+.hero-video-bg {
+	position: absolute;
+	top: 0;
+	left: 0;
+	width: 100%;
+	height: 100%;
+	z-index: 1;
+}
+
+.video-placeholder {
+	width: 100%;
+	height: 100%;
+	background: var(--primary-gradient);
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	font-size: var(--text-4xl);
+	color: var(--text-white);
+	text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+}
+
+.hero-overlay {
+	position: relative;
+	z-index: 2;
+	background: rgba(0, 0, 0, 0.4);
+	width: 100%;
+	height: 100%;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+.hero-content {
+	text-align: center;
+	color: var(--text-white);
+	max-width: 800px;
+	padding: var(--spacing-xl);
+}
+
+.hero-logo h1 {
+	font-size: var(--text-6xl);
+	font-weight: var(--font-black);
+	margin-bottom: var(--spacing-md);
+	text-shadow: 3px 3px 6px rgba(0, 0, 0, 0.5);
+}
+
+.hero-tagline {
+	font-size: var(--text-2xl);
+	margin-bottom: var(--spacing-2xl);
+	font-style: italic;
+	opacity: 0.9;
+}
+
+.hero-cta-buttons {
+	display: flex;
+	gap: var(--spacing-xl);
+	justify-content: center;
+	flex-wrap: wrap;
+}
+
+.cta-btn {
+	padding: 1.2rem 2.5rem;
+	border: none;
+	border-radius: var(--radius-xl);
+	font-size: var(--text-lg);
+	font-weight: var(--font-semibold);
+	cursor: pointer;
+	transition: all var(--transition-normal);
+}
+
+.cta-btn.primary {
+	background: var(--primary-gradient);
+	color: var(--text-white);
+	box-shadow: var(--shadow-primary-lg);
+}
+
+.cta-btn.primary:hover {
+	transform: translateY(-3px);
+	box-shadow: 0 12px 35px rgba(255, 87, 87, 0.6);
+}
+
+.cta-btn.secondary {
+	background: transparent;
+	color: var(--text-white);
+	border: 2px solid var(--text-white);
+}
+
+.cta-btn.secondary:hover {
+	background: var(--text-white);
+	color: var(--primary-color);
+}
+
+/* Responsive Hero Section */
+@media (max-width: 768px) {
+	.hero-logo h1 {
+		font-size: var(--text-5xl);
+	}
+
+	.hero-tagline {
+		font-size: var(--text-xl);
+	}
+
+	.hero-cta-buttons {
+		flex-direction: column;
+		align-items: center;
+	}
+
+	.cta-btn {
+		width: 100%;
+		max-width: 300px;
+	}
+}
