+++ src/styles/sections/LearnSection.css
@@ -1,0 +1,173 @@
+/* Learn Section Styles */
+.learn-section {
+	padding: var(--spacing-4xl) 0;
+	background: var(--bg-white);
+}
+
+.course-filters {
+	display: flex;
+	justify-content: center;
+	gap: var(--spacing-md);
+	margin-bottom: var(--spacing-2xl);
+	flex-wrap: wrap;
+}
+
+.filter-btn {
+	padding: var(--spacing-sm) var(--spacing-lg);
+	border: 2px solid #ddd;
+	background: var(--bg-white);
+	color: var(--text-secondary);
+	border-radius: var(--radius-xl);
+	cursor: pointer;
+	transition: all var(--transition-normal);
+	font-weight: var(--font-semibold);
+}
+
+.filter-btn.active,
+.filter-btn:hover {
+	border-color: var(--primary-color);
+	background: var(--primary-color);
+	color: var(--text-white);
+}
+
+.courses-grid {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(350px, 1fr));
+	gap: var(--spacing-xl);
+	margin-bottom: var(--spacing-3xl);
+}
+
+.course-card {
+	background: var(--bg-white);
+	border-radius: var(--radius-xl);
+	overflow: hidden;
+	box-shadow: var(--shadow-lg);
+	transition: all var(--transition-normal);
+}
+
+.course-card:hover {
+	transform: translateY(-5px);
+	box-shadow: var(--shadow-xl);
+}
+
+.course-video {
+	background: var(--primary-gradient);
+	color: var(--text-white);
+	padding: var(--spacing-2xl);
+	text-align: center;
+	font-size: var(--text-2xl);
+}
+
+.course-info {
+	padding: var(--spacing-xl);
+}
+
+.course-info h3 {
+	margin-bottom: var(--spacing-md);
+	color: var(--text-primary);
+}
+
+.course-meta {
+	display: flex;
+	gap: var(--spacing-md);
+	margin-bottom: var(--spacing-md);
+}
+
+.level {
+	background: var(--primary-color);
+	color: var(--text-white);
+	padding: var(--spacing-xs) var(--spacing-sm);
+	border-radius: var(--radius-md);
+	font-size: var(--text-xs);
+	font-weight: var(--font-semibold);
+}
+
+.teacher {
+	color: var(--text-secondary);
+	font-style: italic;
+}
+
+.course-progress {
+	display: flex;
+	align-items: center;
+	gap: var(--spacing-md);
+}
+
+.progress-bar {
+	flex: 1;
+	height: 8px;
+	background: #ddd;
+	border-radius: 4px;
+	overflow: hidden;
+}
+
+.progress-fill {
+	height: 100%;
+	background: var(--primary-gradient);
+	border-radius: 4px;
+	transition: width var(--transition-normal);
+}
+
+.gamification {
+	background: linear-gradient(135deg, var(--bg-light), var(--bg-white));
+	border-radius: var(--radius-xl);
+	padding: var(--spacing-2xl);
+	text-align: center;
+}
+
+.gamification h3 {
+	margin-bottom: var(--spacing-xl);
+	color: var(--text-primary);
+}
+
+.badges {
+	display: flex;
+	justify-content: center;
+	gap: var(--spacing-md);
+	margin-bottom: var(--spacing-xl);
+	flex-wrap: wrap;
+}
+
+.badge {
+	padding: var(--spacing-md) var(--spacing-lg);
+	border-radius: var(--radius-xl);
+	font-weight: var(--font-semibold);
+	transition: all var(--transition-normal);
+}
+
+.badge.earned {
+	background: var(--primary-gradient);
+	color: var(--text-white);
+	box-shadow: var(--shadow-primary);
+}
+
+.badge:not(.earned) {
+	background: #ddd;
+	color: var(--text-light);
+}
+
+.xp-bar {
+	max-width: 400px;
+	margin: 0 auto;
+}
+
+.xp-bar span {
+	display: block;
+	margin-bottom: var(--spacing-md);
+	font-weight: var(--font-semibold);
+	color: var(--text-primary);
+}
+
+.xp-progress {
+	height: 12px;
+	background: #ddd;
+	border-radius: 6px;
+	overflow: hidden;
+}
+
+.xp-fill {
+	height: 100%;
+	background: var(--primary-gradient);
+	border-radius: 6px;
+	transition: width var(--transition-normal);
+}
