+++ src/utils/constants.js
@@ -1,0 +1,47 @@
+// Application constants
+
+export const NAVIGATION_ITEMS = [
+	{ id: "home", label: "Home", href: "#home" },
+	{ id: "knowledge", label: "Knowledge Library", href: "#knowledge" },
+	{ id: "designer", label: "3D Designer", href: "#designer" },
+	{ id: "learn", label: "Learn & Train", href: "#learn" },
+	{ id: "events", label: "Live Events", href: "#events" },
+	{ id: "community", label: "Community", href: "#community" },
+	{ id: "competitions", label: "Competitions", href: "#competitions" },
+];
+
+export const KNOWLEDGE_TABS = [
+	{ id: "history", label: "Lịch sử" },
+	{ id: "types", label: "Các loại hình" },
+	{ id: "regions", label: "Theo vùng miền" },
+	{ id: "unesco", label: "UNESCO Di sản" },
+];
+
+export const REGIONS = [
+	{ id: "all", label: "All Regions" },
+	{ id: "north", label: "North Vietnam" },
+	{ id: "central", label: "Central Vietnam" },
+	{ id: "south", label: "South Vietnam" },
+];
+
+export const COMMUNITY_TABS = [
+	{ id: "marketplace", label: "Marketplace" },
+	{ id: "stories", label: "Stories" },
+	{ id: "wall", label: "Community Wall" },
+];
+
+export const COURSE_FILTERS = [
+	{ id: "all", label: "All" },
+	{ id: "basic", label: "Basic" },
+	{ id: "advanced", label: "Advanced" },
+	{ id: "region", label: "By Region" },
+	{ id: "free", label: "Free" },
+	{ id: "certified", label: "Certified" },
+];
+
+export const MARKETPLACE_CATEGORIES = [
+	{ id: "lion-heads", label: "Lion Heads" },
+	{ id: "costumes", label: "Costumes" },
+	{ id: "accessories", label: "Accessories" },
+	{ id: "performance", label: "Performance Booking" },
+];
